# Upgrade reqwest to v0.10.0 for better binary size and build time