--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,5 +17,9 @@ clap = "2.33.0"
 html5ever = "0.24.1"
 lazy_static = "1.4.0"
 regex = "1.3.1"
-reqwest = "0.9.20"
 url = "2.1.0"
+
+[dependencies.reqwest]
+version = "0.10.*"
+default-features = false
+features = ["default-tls", "blocking", "gzip"]