--- a/src/http.rs
+++ b/src/http.rs
@@ -1,6 +1,6 @@
 use crate::utils::{clean_url, data_to_dataurl, is_data_url};
+use reqwest::blocking::Client;
 use reqwest::header::CONTENT_TYPE;
-use reqwest::Client;
 use std::collections::HashMap;
 
 pub fn retrieve_asset(
@@ -26,17 +26,17 @@ pub fn retrieve_asset(
         } else {
             // url not in cache, we request it
             let mut response = client.get(url).send()?;
+            let res_url = response.url().to_string();
 
             if !opt_silent {
-                let res_url = response.url().as_str();
                 if url == res_url {
                     eprintln!("{}", &url);
                 } else {
-                    eprintln!("{} -> {}", &url, res_url);
+                    eprintln!("{} -> {}", &url, &res_url);
                 }
             }
 
-            let new_cache_key = clean_url(response.url());
+            let new_cache_key = clean_url(&res_url);
 
             if as_dataurl {
                 // Convert response into a byte array
@@ -56,12 +56,12 @@ pub fn retrieve_asset(
                 let dataurl = data_to_dataurl(&mimetype, &data);
                 // insert in cache
                 cache.insert(new_cache_key, dataurl.clone());
-                Ok((dataurl, response.url().to_string()))
+                Ok((dataurl, res_url))
             } else {
                 let content = response.text().unwrap();
                 // insert in cache
                 cache.insert(new_cache_key, content.clone());
-                Ok((content, response.url().to_string()))
+                Ok((content, res_url))
             }
         }
     }