--- a/src/tests/http.rs
+++ b/src/tests/http.rs
@@ -1,9 +1,11 @@
 use crate::http::retrieve_asset;
+use reqwest::blocking::Client;
 use std::collections::HashMap;
+
 #[test]
 fn test_retrieve_asset() {
     let cache = &mut HashMap::new();
-    let client = reqwest::Client::new();
+    let client = Client::new();
     let (data, final_url) =
         retrieve_asset(cache, &client, "data:text/html;base64,...", true, "", false).unwrap();
     assert_eq!(&data, "data:text/html;base64,...");