I updated `reqwest` dependency from v0.9 to [v0.10](https://github.com/seanmonstar/reqwest/releases/tag/v0.10.0).

Effects are as follows:

### Before

- **Compile targets**: 220
- **Build time**: `cargo build --release  1264.95s user 39.72s system 335% cpu 6:29.14 total`
- **Binary size**: 6578568 bytes

### After

- **Compile targets**: 170
- **Build time**: `cargo build --release  1130.64s user 32.15s system 359% cpu 5:23.69 total`
- **Binary size**: 6107088 bytes

### Differences

- **Compile targets**: 1.29x smaller
- **Build time**: 1.23x faster
- **Binary size**: 1.07x smaller

reqwest 0.10 now separates its functionalities into several features. I picked necessary features (all features list can be found [here](https://github.com/seanmonstar/reqwest/blob/b159963f6c19d8607a38f1a17c45b38fa8b4d4d2/Cargo.toml#L25-L48)) so number of dependencies was reduced. As the result build time and binary size are improved.