@rhysd 2020-01-08
Thank you for your detailed review.